requirements.txt
opencv-python
Pillow
ultralytics
numpy
matplotlib
colorgram.py
transformers
torch
